requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
openai==1.3.5
Pillow==10.1.0
python-pptx==0.6.23
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
django-environ==0.11.2
requests==2.31.0
python-multipart==0.0.6
